# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LEVH API",
    "description": "Local-first memory layer for AI agents and humans",
    "version": "2.32.0"
  },
  "paths": {
...
    "/api/v1/attachments/upload": {
      "post": {
        "summary": "Upload Attachment",
        "description": "Store an uploaded file locally and return the path to attach from.",
        "operationId": "upload_attachment_api_attachments_upload_post_v1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachmentUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadedFileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memories/{memory_id}/attachments": {
      "post": {
        "summary": "Attach File",
        "description": "Attach a local file to a memory by reference (path + sha256), with\noptional derived text (OCR/transcript/caption) that recall actually\nsearches over.",
        "operationId": "attach_file_api_memories__memory_id__attachments_post_v1",
        "parameters": [
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachFileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Memory Attachments",
        "operationId": "list_memory_attachments_api_memories__memory_id__attachments_get_v1",
        "parameters": [
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attachments/{attachment_id}/verify": {
      "post": {
        "summary": "Verify Attachment",
        "description": "Re-check the file against what was recorded at attach time. A missing\nor changed file raises a conflict candidate rather than silently altering\nthe memory \u2014 see /api/conflicts.",
        "operationId": "verify_attachment_api_attachments__attachment_id__verify_post_v1",
        "parameters": [
          {
            "name": "attachment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attachment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentVerifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/attachments/{attachment_id}": {
      "delete": {
        "summary": "Delete Attachment",
        "operationId": "delete_attachment_api_attachments__attachment_id__delete_v1",
        "parameters": [
          {
            "name": "attachment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attachment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/conflicts": {
      "get": {
        "summary": "List Conflicts",
        "description": "List conflict candidates, optionally filtered by status. Pass an empty\nstatus to list every status.",
        "operationId": "list_conflicts_api_conflicts_get_v1",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "open",
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/conflicts/{conflict_id}/review": {
      "post": {
        "summary": "Review Conflict",
        "description": "Apply a human review decision to a conflict candidate. ``conflict_id``\nmay contain ``|`` so the path converter is used.",
        "operationId": "review_conflict_api_conflicts__conflict_id__review_post_v1",
        "parameters": [
          {
            "name": "conflict_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conflict Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConflictReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttachFileRequest": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "derived_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Derived Text"
          },
          "derived_by": {
            "type": "string",
            "title": "Derived By",
            "default": "manual"
          }
        },
        "type": "object",
        "required": [
          "path"
        ],
        "title": "AttachFileRequest"
      },
      "AttachmentDeleteResponse": {
        "properties": {
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          }
        },
        "type": "object",
        "required": [
          "deleted"
        ],
        "title": "AttachmentDeleteResponse"
      },
      "AttachmentListResponse": {
        "properties": {
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentOut"
            },
            "type": "array",
            "title": "Attachments"
          }
        },
        "type": "object",
        "required": [
          "attachments"
        ],
        "title": "AttachmentListResponse",
        "description": "``GET /api/memories/{id}/attachments`` - every file attached to one\nmemory."
      },
      "AttachmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "memory_id": {
            "type": "string",
            "title": "Memory Id"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "mime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "derived_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Derived Text"
          },
          "derived_by": {
            "type": "string",
            "title": "Derived By",
            "default": "none"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ok"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "verified_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "memory_id",
          "path",
          "sha256",
          "size",
          "created_at"
        ],
        "title": "AttachmentOut",
        "description": "A file attached to a memory (the ``attachments`` table row)."
      },
      "AttachmentUploadRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content_b64": {
            "type": "string",
            "title": "Content B64"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "content_b64"
        ],
        "title": "AttachmentUploadRequest"
      },
      "AttachmentVerifyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "memory_id": {
            "type": "string",
            "title": "Memory Id"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "mime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "derived_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Derived Text"
          },
          "derived_by": {
            "type": "string",
            "title": "Derived By",
            "default": "none"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ok"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "verified_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "memory_id",
          "path",
          "sha256",
          "size",
          "created_at"
        ],
        "title": "AttachmentVerifyResponse",
        "description": "``POST /api/attachments/{id}/verify`` - the row with its post-check\nstatus (``ok`` / ``missing`` / ``changed``)."
      },
...
      "ConflictListResponse": {
        "properties": {
          "conflicts": {
            "items": {
              "$ref": "#/components/schemas/ConflictOut"
            },
            "type": "array",
            "title": "Conflicts"
          }
        },
        "type": "object",
        "title": "ConflictListResponse"
      },
      "ConflictOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "memory_id_a": {
            "type": "string",
            "title": "Memory Id A"
          },
          "memory_id_b": {
            "type": "string",
            "title": "Memory Id B"
          },
          "signal_type": {
            "type": "string",
            "title": "Signal Type"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "shared_entities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Shared Entities"
          },
          "explanation": {
            "additionalProperties": true,
            "type": "object",
            "title": "Explanation"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "memory_id_a",
          "memory_id_b",
          "signal_type",
          "confidence",
          "status"
        ],
        "title": "ConflictOut",
        "description": "A conflict-candidate row, JSON columns parsed."
      },
      "ConflictReviewRequest": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "ConflictReviewRequest"
      },
      "ConflictReviewResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "conflict": {
            "$ref": "#/components/schemas/ConflictOut"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "action",
          "conflict"
        ],
        "title": "ConflictReviewResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UploadedFileResponse": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "bytes": {
            "type": "integer",
            "title": "Bytes"
          }
        },
        "type": "object",
        "required": [
          "path",
          "filename",
          "bytes"
        ],
        "title": "UploadedFileResponse",
        "description": "Shared by the upload endpoints: where the bytes landed on this\nmachine, and the display name the caller chose (never the stored path)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}